Some general refactoring, housecleaning, code cleanup, and renaming as well as add a few tests that I forgot to add for the controller and use cases.
